pyproject.toml
[project]
name = "be"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.18.4",
    "fastapi>=0.135.3",
    "firebase-admin>=7.4.0",
    "google-genai>=1.73.1",
    "pgvector>=0.4.2",
    "psycopg[binary,pool]>=3.3.3",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sentence-transformers>=5.4.1",
    "sqlmodel>=0.0.38",
    "uvicorn[standard]>=0.44.0",
] 
